Move microgrid config to new data namespace
The new namespace will contain data-ingestion related modules and may be
split into a separate repo if needed.
