+++ src/frequenz/data/microgrid/__init__.py
@@ -1,0 +1,10 @@
+# License: MIT
+# Copyright © 2025 Frequenz Energy-as-a-Service GmbH
+
+"""Initialize the microgrid data module."""
+
+from .config import MicrogridConfig
+
+__all__ = [
+    "MicrogridConfig",
+]
